requirements.txt
tensorflow==1.12.0
tqdm
numpy
pillow
scipy
opencv-python
